requirements.txt
# ===========================
# OpenLabel Backend Dependencies
# ===========================

# Web Framework
fastapi>=0.115.0
uvicorn[standard]>=0.34.0

# Database — Supabase
supabase>=2.12.0

# Data Validation
pydantic>=2.10.0
pydantic-settings>=2.7.0

# AI — Google Gemini (Member 3 uses legacy `google-generativeai`)
google-generativeai>=0.8.0

# OCR + OSINT live context
google-cloud-vision>=3.0.0
requests>=2.28.0
Pillow>=10.0.0
pytesseract>=0.3.10

# HTTP Client (async)
httpx>=0.28.0

# Web Scraping
beautifulsoup4>=4.12.0
lxml>=5.3.0

# Utilities
python-dotenv>=1.0.0
python-multipart>=0.0.18
